+++ extra-tests/alt-store/flake.nix
@@ -5,9 +5,10 @@
     flake-utils.url = "github:numtide/flake-utils";
   };
 
-  outputs = { flake-utils, crane, ... }: flake-utils.lib.eachDefaultSystem (system:
+  outputs = { nixpkgs, flake-utils, crane, ... }: flake-utils.lib.eachDefaultSystem (system:
     let
-      craneLib = crane.lib.${system};
+      pkgs = nixpkgs.legacyPackages.${system};
+      craneLib = crane.mkLib pkgs;
     in
     {
       # https://github.com/ipetkov/crane/issues/446