+++ extra-tests/dummy-does-not-depend-on-flake-source-via-self/flake.nix
@@ -5,9 +5,13 @@
     flake-utils.url = "github:numtide/flake-utils";
   };
 
-  outputs = { self, flake-utils, crane, ... }: flake-utils.lib.eachDefaultSystem
-    (system: {
-      packages.dummy = crane.lib.${system}.mkDummySrc {
+  outputs = { self, nixpkgs, flake-utils, crane, ... }: flake-utils.lib.eachDefaultSystem (system:
+    let
+      pkgs = nixpkgs.legacyPackages.${system};
+      craneLib = crane.mkLib pkgs;
+    in
+    {
+      packages.dummy = craneLib.mkDummySrc {
         src = self;
       };
     });